some change to datasets

+++ utils/create_datasets.py
@@ -322,8 +322,8 @@ class SumDatasets(Dataset):
         features_1, features_2 = self.features_1[index], self.features_2[index]
         features_3, features_4 = self.features_3[index], self.features_4[index]
         # convert to torch datasets
-        features_1, features_2 = torch.from_numpy(features_1).long(), torch.from_numpy(features_2).long()
-        features_3, features_4 = torch.from_numpy(features_3).long(), torch.from_numpy(features_4).long()
+        features_1, features_2 = torch.from_numpy(features_1), torch.from_numpy(features_2)
+        features_3, features_4 = torch.from_numpy(features_3), torch.from_numpy(features_4)
         if len(features_1.size()) == 1: # if we only have one item, write it to
             features_1.unsqueeze_(0), features_3.unsqueeze_(0), features_4.unsqueeze_(0)
         return self.transform(features_1, features_2, features_3, features_4)
